requirements.txt
pandas
numpy
streamlit
scikit-learn
pymongo
pymongo[srv]==3.12
dotenv
pyyaml
imblearn
xgboost
fastapi
uvicorn